requirements.txt
# Core Framework
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0
pandas>=2.1.0
numpy>=1.24.0

# Industrial Automation & Control
pyModbusTCP>=0.2.0
opcua>=0.98.13
python-control>=0.9.4

# Signal Processing & Analysis
opencv-python>=4.8.0
scipy>=1.11.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Database & Time Series
influxdb-client>=1.38.0
sqlite3
redis>=4.6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
bokeh>=3.2.0

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
asyncio-mqtt>=0.13.0

# Industrial Protocols
snap7>=1.3.0
pycomm3>=1.2.14

# Web Development
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Docker & Deployment
docker>=6.1.0
python-multipart>=0.0.6

# Mathematical & Engineering
sympy>=1.12.0
control>=0.9.4 